+++ src/main/java/net/lopymine/ip/resourcepack/ResourcePackParticleConfigsManager.java
@@ -34,7 +34,7 @@ public class ResourcePackParticleConfigsManager {
 			try (InputStream inputStream = resource.open(); BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream, StandardCharsets.UTF_8))) {
 				ParticleConfig config = ParticleConfig.CODEC.decode(JsonOps.INSTANCE, JsonParser.parseReader(reader))/*? if >=1.20.5 {*/.getOrThrow()/*?} else {*//*.getOrThrow(false, InventoryParticlesClient.LOGGER::error)*//*?}*/.getFirst();
 				for (ParticleHolder holder : config.getHolders()) {
-					Item item = holder.getItem().getItem();
+					Item item = holder.getItemOrTag().getItem();
 					ParticleSpawner spawner = holder.create(config::createParticle);
 					registerItemSpawner(id, item, holder, spawner);
 				}