+++ src/main/java/net/lopymine/ip/tag/ItemTagProvider.java
@@ -1,0 +1,5 @@
+package net.lopymine.ip.tag;
+
+public interface ItemTagProvider {
+
+}